pyproject.toml
[tool.poetry]
name = "simple-searcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.10.5"
elasticsearch = "^8.15.0"
black = "^24.8.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.34"
aiohttp-swagger = "^1.0.16"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
